Add more test cases, refactor
